api/requirements.txt
Django==1.8.5
djangorestframework==3.2.4
django-cors-headers
psycopg2
requests